requirements.txt
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.5.4
